Remove websocket notification code
The notification system has been migrated to SSE. Remove the now-unused
WebSocket handler, tests, and the coder/websocket dependency. The
notifyserver.Server now exposes a single Handler() that serves SSE.
