+++ internal/server/notifyserver/sse_test.go
@@ -16,6 +16,23 @@ import (
 	"gotest.tools/v3/assert"
 )
 
+// allowOrgResolver returns a mock that allows the given user to subscribe
+// to any of the configured orgs and rejects everything else.
+func allowOrgResolver(userID string, allow ...int64) *OrgResolverMock {
+	allowed := make(map[int64]bool, len(allow))
+	for _, id := range allow {
+		allowed[id] = true
+	}
+	return &OrgResolverMock{
+		ResolveOrgFunc: func(ctx context.Context, callerID string, orgID int64) (int64, error) {
+			if callerID != userID || !allowed[orgID] {
+				return 0, fmt.Errorf("user %s not a member of org %d", callerID, orgID)
+			}
+			return orgID, nil
+		},
+	}
+}
+
 func TestSSE(t *testing.T) {
 	t.Parallel()
 
@@ -26,7 +43,7 @@ func TestSSE(t *testing.T) {
 		OrgResolver: allowOrgResolver("u", orgID),
 	})
 
-	ts := httptest.NewServer(apiauth.WithTestUser(srv.SSEHandler(), &apiauth.UserInfo{ID: "u", OrgID: orgID}))
+	ts := httptest.NewServer(apiauth.WithTestUser(srv.Handler(), &apiauth.UserInfo{ID: "u", OrgID: orgID}))
 	defer ts.Close()
 
 	ctx, cancel := context.WithTimeout(t.Context(), 10*time.Second)
@@ -82,7 +99,7 @@ func TestSSE_OrgIsolation(t *testing.T) {
 		OrgResolver: allowOrgResolver("u", orgB),
 	})
 
-	ts := httptest.NewServer(apiauth.WithTestUser(srv.SSEHandler(), &apiauth.UserInfo{ID: "u", OrgID: orgB}))
+	ts := httptest.NewServer(apiauth.WithTestUser(srv.Handler(), &apiauth.UserInfo{ID: "u", OrgID: orgB}))
 	defer ts.Close()
 
 	ctx, cancel := context.WithTimeout(t.Context(), 10*time.Second)
